Move check-related types to own file
